requirements.txt
streamlit
langchain
openai
python-dotenv
langchain-openai
